move publish func to a separate file
