+++ openml/__init__.py
@@ -18,9 +18,6 @@ In particular, this module implements a python interface for the
 # License: BSD 3-Clause
 from __future__ import annotations
 
-import contextlib
-from typing import Any, Sequence
-
 from . import (
     _api_calls,
     config,
@@ -36,10 +33,10 @@ from . import (
     utils,
 )
 from .__version__ import __version__
-from .base import OpenMLBase
 from .datasets import OpenMLDataFeature, OpenMLDataset
 from .evaluations import OpenMLEvaluation
 from .flows import OpenMLFlow
+from .publish import publish
 from .runs import OpenMLRun
 from .setups import OpenMLParameter, OpenMLSetup
 from .study import OpenMLBenchmarkSuite, OpenMLStudy
@@ -52,122 +49,6 @@ from .tasks import (
     OpenMLSupervisedTask,
     OpenMLTask,
 )
-
-
-def publish(obj: Any, *, name: str | None = None, tags: Sequence[str] | None = None) -> Any:
-    """Publish a common object (flow/model/run/dataset) with minimal friction.
-
-    This function provides a unified entry point for publishing various OpenML objects.
-    It automatically detects the object type and routes to the appropriate publishing
-    mechanism:
-
-    - For OpenML objects (``OpenMLDataset``, ``OpenMLFlow``, ``OpenMLRun``, etc.),
-      it directly calls their ``publish()`` method.
-    - For external models (e.g., scikit-learn estimators), it uses registered
-      extensions to convert them to ``OpenMLFlow`` objects before publishing.
-
-    Parameters
-    ----------
-    obj : Any
-        The object to publish. Can be:
-        - An OpenML object (OpenMLDataset, OpenMLFlow, OpenMLRun, OpenMLTask)
-        - A machine learning model from a supported framework (e.g., scikit-learn)
-    name : str, optional
-        Override the default name for the published object.
-        If not provided, uses the object's default naming convention.
-    tags : Sequence[str], optional
-        Additional tags to attach to the published object.
-        Will be merged with any existing tags, removing duplicates while
-        preserving order.
-
-    Returns
-    -------
-    Any
-        The published object (typically with updated ID and metadata).
-
-    Raises
-    ------
-    ValueError
-        If no extension is registered to handle the provided model type.
-
-    Examples
-    --------
-    Publishing an OpenML dataset:
-
-    >>> dataset = openml.datasets.get_dataset(61)
-    >>> openml.publish(dataset, tags=["example"])
-
-    Publishing a scikit-learn model:
-
-    >>> from sklearn.tree import DecisionTreeClassifier
-    >>> clf = DecisionTreeClassifier(max_depth=5)
-    >>> openml.publish(clf, name="MyDecisionTree", tags=["tutorial"])
-
-    Publishing an OpenML flow directly:
-
-    >>> flow = openml.flows.OpenMLFlow(...)
-    >>> openml.publish(flow)
-
-    Publishing an OpenML run (after execution with predictions):
-
-    >>> run = openml.runs.OpenMLRun(
-    ...     task_id=1, flow_id=100, dataset_id=61,
-    ...     data_content=predictions  # predictions from model evaluation
-    ... )
-    >>> openml.publish(run, tags=["experiment"])
-
-    Notes
-    -----
-    For external models (e.g., scikit-learn), the corresponding extension must be
-    installed (e.g., ``openml-sklearn``). The extension will be automatically imported
-    if available.
-    """
-    # Case 1: Object is already an OpenML entity
-    if isinstance(obj, OpenMLBase):
-        if tags is not None and hasattr(obj, "tags"):
-            existing = list(getattr(obj, "tags", []) or [])
-            merged = list(dict.fromkeys([*existing, *tags]))
-            obj.tags = merged
-        if name is not None and hasattr(obj, "name"):
-            obj.name = name
-        return obj.publish()
-
-    # Case 2: Object is an external model - use extension registry
-    # Attempt to auto-import common extensions
-    _ensure_extension_imported(obj)
-
-    extension = extensions.functions.get_extension_by_model(obj, raise_if_no_extension=True)
-    if extension is None:  # Defensive check (should not occur with raise_if_no_extension=True)
-        raise ValueError("No extension registered to handle the provided object.")
-    flow = extension.model_to_flow(obj)
-
-    if name is not None:
-        flow.name = name
-
-    if tags is not None:
-        existing_tags = list(getattr(flow, "tags", []) or [])
-        flow.tags = list(dict.fromkeys([*existing_tags, *tags]))
-
-    return flow.publish()
-
-
-def _ensure_extension_imported(obj: Any) -> None:
-    """Attempt to import the appropriate extension for common frameworks.
-
-    This is a convenience helper to automatically import extensions for
-    well-known frameworks, reducing friction for users.
-
-    Parameters
-    ----------
-    obj : Any
-        The object to check.
-    """
-    obj_module = type(obj).__module__
-
-    # Check for scikit-learn models
-    if obj_module.startswith("sklearn"):
-        with contextlib.suppress(ImportError):
-            import openml_sklearn  # noqa: F401
 
 
 def populate_cache(
